Move "deep_update" helper to tools/pytest_fixtures/webdriver.py
Issue: #59157
